requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.7
python-decouple==3.8
PyJWT==2.8.0
django-oauth-toolkit==1.7.1
django-ratelimit==4.1.0
celery==5.3.4
redis==5.0.1
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.25.2
joblib==1.3.2
drf-yasg==1.21.7
django-extensions==3.2.3
Pillow==10.1.0
python-dotenv==1.0.0
gunicorn==21.2.0
whitenoise==6.6.0